+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutClient.java
@@ -85,7 +85,7 @@ public class PayPalNativeCheckoutClient {
      */
     @Deprecated
     public void tokenizePayPalAccount(@NonNull final FragmentActivity activity, @NonNull final PayPalNativeRequest payPalRequest) throws Exception {
-        braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.started", payPalContextId);
+        braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.started", payPalContextId, null);
         // NEXT_MAJOR_VERSION: remove tokenizePayPalAccount method and refactor tests to center
         // around launchNativeCheckout in the future. Keeping the tests as they are for now allows
         // us to maintain test coverage across both the tokenizePayPalAccount and launchNativeCheckout methods
@@ -94,7 +94,7 @@ public class PayPalNativeCheckoutClient {
         if (isCheckoutRequest || isVaultRequest) {
             launchNativeCheckout(activity, payPalRequest);
         } else {
-            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.invalid-request.failed", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.invalid-request.failed", payPalContextId, null);
             String message = "Unsupported request type. Please use either a "
                     + "PayPalNativeCheckoutRequest or a PayPalNativeCheckoutVaultRequest.";
             throw new Exception(message);
@@ -111,15 +111,15 @@ public class PayPalNativeCheckoutClient {
      * @param payPalRequest a {@link PayPalNativeRequest} used to customize the request.
      */
     public void launchNativeCheckout(@NonNull final FragmentActivity activity, @NonNull final PayPalNativeRequest payPalRequest) {
-        braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.started", payPalContextId);
+        braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.started", payPalContextId, null);
         if (payPalRequest instanceof PayPalNativeCheckoutRequest) {
             sendCheckoutRequest(activity, (PayPalNativeCheckoutRequest) payPalRequest);
-            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.succeeded", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.succeeded", payPalContextId, null);
         } else if (payPalRequest instanceof PayPalNativeCheckoutVaultRequest) {
             sendVaultRequest(activity, (PayPalNativeCheckoutVaultRequest) payPalRequest);
-            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.succeeded", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.succeeded", payPalContextId, null);
         } else if (listener != null) {
-            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.invalid-request.failed", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("paypal-native.tokenize.invalid-request.failed", payPalContextId, null);
             String message = "Unsupported request type. Please use either a "
                     + "PayPalNativeCheckoutRequest or a PayPalNativeCheckoutVaultRequest.";
             listener.onPayPalFailure(new BraintreeException(message));
@@ -127,9 +127,9 @@ public class PayPalNativeCheckoutClient {
     }
 
     private void sendCheckoutRequest(final FragmentActivity activity, final PayPalNativeCheckoutRequest payPalCheckoutRequest) {
-        braintreeClient.sendAnalyticsEvent("paypal-native.single-payment.selected", payPalContextId);
+        braintreeClient.sendAnalyticsEvent("paypal-native.single-payment.selected", payPalContextId, null);
         if (payPalCheckoutRequest.getShouldOfferPayLater()) {
-            braintreeClient.sendAnalyticsEvent("paypal-native.single-payment.paylater.offered", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("paypal-native.single-payment.paylater.offered", payPalContextId, null);
         }
 
         braintreeClient.getConfiguration((configuration, error) -> {
@@ -138,9 +138,9 @@ public class PayPalNativeCheckoutClient {
     }
 
     private void sendVaultRequest(final FragmentActivity activity, final PayPalNativeCheckoutVaultRequest payPalVaultRequest) {
-        braintreeClient.sendAnalyticsEvent("paypal-native.billing-agreement.selected", payPalContextId);
+        braintreeClient.sendAnalyticsEvent("paypal-native.billing-agreement.selected", payPalContextId, null);
         if (payPalVaultRequest.getShouldOfferCredit()) {
-            braintreeClient.sendAnalyticsEvent("paypal-native.billing-agreement.credit.offered", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("paypal-native.billing-agreement.credit.offered", payPalContextId, null);
         }
 
         braintreeClient.getConfiguration((configuration, error) -> {
@@ -160,7 +160,7 @@ public class PayPalNativeCheckoutClient {
                     payPalContextId = pairingId;
                 }
                 String analyticsPrefix = payPalRequest instanceof PayPalNativeCheckoutVaultRequest ? "billing-agreement" : "single-payment";
-                braintreeClient.sendAnalyticsEvent(String.format("paypal-native.%s.started", analyticsPrefix), payPalContextId);
+                braintreeClient.sendAnalyticsEvent(String.format("paypal-native.%s.started", analyticsPrefix), payPalContextId, null);
 
                 Environment environment;
                 if ("sandbox".equals(configuration.getEnvironment())) {
@@ -213,10 +213,10 @@ public class PayPalNativeCheckoutClient {
                 PayPalCheckout.startCheckout(createOrderActions -> {
                     if (payPalRequest instanceof PayPalNativeCheckoutRequest) {
                         createOrderActions.set(payPalResponse.getPairingId());
-                        braintreeClient.sendAnalyticsEvent("paypal-native.single-payment.succeeded", payPalContextId);
+                        braintreeClient.sendAnalyticsEvent("paypal-native.single-payment.succeeded", payPalContextId, null);
                     } else if (payPalRequest instanceof PayPalNativeCheckoutVaultRequest) {
                         createOrderActions.setBillingAgreementId(payPalResponse.getPairingId());
-                        braintreeClient.sendAnalyticsEvent("paypal-native.billing-agreement.succeeded", payPalContextId);
+                        braintreeClient.sendAnalyticsEvent("paypal-native.billing-agreement.succeeded", payPalContextId, null);
                     }
                 }, payPalRequest.hasUserLocationConsent());
             } else {
@@ -232,25 +232,25 @@ public class PayPalNativeCheckoutClient {
     ) {
         PayPalCheckout.registerCallbacks(
                 approval -> {
-                    braintreeClient.sendAnalyticsEvent("paypal-native.on-approve.started", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("paypal-native.on-approve.started", payPalContextId, null);
                     PayPalNativeCheckoutAccount payPalAccount = setupAccount(payPalRequest, approval.getData());
                     internalPayPalClient.tokenize(payPalAccount, (payPalAccountNonce, error) -> {
                         if (payPalAccountNonce != null) {
-                            braintreeClient.sendAnalyticsEvent("paypal-native.on-approve.succeeded", payPalContextId);
+                            braintreeClient.sendAnalyticsEvent("paypal-native.on-approve.succeeded", payPalContextId, null);
                             listener.onPayPalSuccess(payPalAccountNonce);
                         } else {
-                            braintreeClient.sendAnalyticsEvent("paypal-native.on-approve.failed", payPalContextId);
+                            braintreeClient.sendAnalyticsEvent("paypal-native.on-approve.failed", payPalContextId, null);
                             listener.onPayPalFailure(new Exception("PaypalAccountNonce is null"));
                         }
                     });
                 },
                 null,
                 () -> {
-                    braintreeClient.sendAnalyticsEvent("paypal-native.canceled", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("paypal-native.canceled", payPalContextId, null);
                     listener.onPayPalFailure(new Exception("User has canceled"));
                 },
                 errorInfo -> {
-                    braintreeClient.sendAnalyticsEvent("paypal-native.on-error.failed", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("paypal-native.on-error.failed", payPalContextId, null);
                     listener.onPayPalFailure(new Exception(errorInfo.getError().getMessage()));
                 }
         );