+++ Venmo/src/main/java/com/braintreepayments/api/VenmoClient.java
@@ -47,6 +47,11 @@ public class VenmoClient {
      */
     private String payPalContextId = null;
 
+    /**
+     * Used for sending the type of flow, universal vs deeplink to FPTI
+     */
+    private String linkType = null;
+
     @VisibleForTesting
     VenmoLifecycleObserver observer;
 
@@ -126,7 +131,7 @@ public class VenmoClient {
      * @param activity used to open the Venmo's Google Play Store
      */
     public void showVenmoInGooglePlayStore(@NonNull FragmentActivity activity) {
-        braintreeClient.sendAnalyticsEvent("android.pay-with-venmo.app-store.invoked", payPalContextId);
+        braintreeClient.sendAnalyticsEvent("android.pay-with-venmo.app-store.invoked", payPalContextId, linkType);
         Intent intent = new Intent(Intent.ACTION_VIEW);
         intent.setData(Uri.parse(
                 "https://play.google.com/store/apps/details?id=" + VENMO_PACKAGE_NAME));
@@ -165,13 +170,14 @@ public class VenmoClient {
      */
     @Deprecated
     public void tokenizeVenmoAccount(@NonNull final FragmentActivity activity, @NonNull final VenmoRequest request, @NonNull final VenmoTokenizeAccountCallback callback) {
-        braintreeClient.sendAnalyticsEvent("pay-with-venmo.selected", payPalContextId);
+        linkType = request.getFallbackToWeb() ? "universal" : "deeplink";
+        braintreeClient.sendAnalyticsEvent("pay-with-venmo.selected", payPalContextId, linkType);
         braintreeClient.getConfiguration(new ConfigurationCallback() {
             @Override
             public void onResult(@Nullable final Configuration configuration, @Nullable Exception error) {
                 if (configuration == null) {
                     callback.onResult(error);
-                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId, linkType);
                     return;
                 }
 
@@ -188,14 +194,14 @@ public class VenmoClient {
 
                 if (exceptionMessage != null) {
                     callback.onResult(new AppSwitchNotAvailableException(exceptionMessage));
-                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId, linkType);
                     return;
                 }
 
                 // Merchants are not allowed to collect user addresses unless ECD (Enriched Customer Data) is enabled on the BT Control Panel.
                 if ((request.getCollectCustomerShippingAddress() || request.getCollectCustomerBillingAddress()) && !configuration.getVenmoEnrichedCustomerDataEnabled()) {
                     callback.onResult(new BraintreeException("Cannot collect customer data when ECD is disabled. Enable this feature in the Control Panel to collect this data."));
-                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId, linkType);
                     return;
                 }
 
@@ -226,7 +232,7 @@ public class VenmoClient {
                             });
                         } else {
                             callback.onResult(exception);
-                            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId);
+                            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failed", payPalContextId, linkType);
                         }
                     }
                 });
@@ -251,7 +257,7 @@ public class VenmoClient {
                 try {
                     startAppLinkFlow(activity, intentData);
                 } catch (JSONException | BrowserSwitchException exception) {
-                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure");
+                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure", payPalContextId, linkType);
                     deliverVenmoFailure(exception);
                 }
             } else {
@@ -261,12 +267,12 @@ public class VenmoClient {
             Intent launchIntent = getLaunchIntent(configuration, venmoProfileId, paymentContextId);
             activity.startActivityForResult(launchIntent, BraintreeRequestCodes.VENMO);
         }
-        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.started", payPalContextId);
+        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.started", payPalContextId, linkType);
     }
 
     void onVenmoResult(final VenmoResult venmoResult) {
         if (venmoResult.getError() == null) {
-            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.success", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.success", payPalContextId, linkType);
 
             braintreeClient.getAuthorization(new AuthorizationCallback() {
                 @Override
@@ -293,11 +299,11 @@ public class VenmoClient {
                                                 }
                                             });
                                         } else {
-                                            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId);
+                                            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId, linkType);
                                             deliverVenmoSuccess(nonce);
                                         }
                                     } else {
-                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId);
+                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId, linkType);
                                         deliverVenmoFailure(error);
                                     }
                                 }
@@ -332,7 +338,7 @@ public class VenmoClient {
 
         } else if (venmoResult.getError() != null) {
             if (venmoResult.getError() instanceof UserCanceledException) {
-                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.canceled", payPalContextId);
+                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.canceled", payPalContextId, linkType);
             }
             deliverVenmoFailure(venmoResult.getError());
         }
@@ -362,7 +368,7 @@ public class VenmoClient {
      */
     public void onActivityResult(@NonNull final Context context, int resultCode, @Nullable final Intent data, @NonNull final VenmoOnActivityResultCallback callback) {
         if (resultCode == AppCompatActivity.RESULT_OK) {
-            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.success", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.success", payPalContextId, linkType);
 
             braintreeClient.getAuthorization(new AuthorizationCallback() {
                 @Override
@@ -380,12 +386,12 @@ public class VenmoClient {
                                         if (shouldVault && isClientTokenAuth) {
                                             vaultVenmoAccountNonce(nonce.getString(), callback);
                                         } else {
-                                            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId);
+                                            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId, linkType);
                                             callback.onResult(nonce, null);
                                         }
 
                                     } else {
-                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId);
+                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.failure", payPalContextId, linkType);
                                         callback.onResult(null, error);
                                     }
                                 }
@@ -409,7 +415,7 @@ public class VenmoClient {
             });
 
         } else if (resultCode == AppCompatActivity.RESULT_CANCELED) {
-            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.canceled", payPalContextId);
+            braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-switch.canceled", payPalContextId, linkType);
             callback.onResult(null, new UserCanceledException("User canceled Venmo."));
         }
     }
@@ -419,9 +425,9 @@ public class VenmoClient {
             @Override
             public void onResult(@Nullable VenmoAccountNonce venmoAccountNonce, @Nullable Exception error) {
                 if (venmoAccountNonce != null) {
-                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.vault.success", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.vault.success", payPalContextId, linkType);
                 } else {
-                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.vault.failed", payPalContextId);
+                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.vault.failed", payPalContextId, linkType);
                 }
                 callback.onResult(venmoAccountNonce, error);
             }
@@ -469,7 +475,7 @@ public class VenmoClient {
         int result = browserSwitchResult.getStatus();
         switch (result) {
             case BrowserSwitchStatus.CANCELED:
-                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.canceled");
+                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.canceled", payPalContextId, linkType);
                 callback.onResult(null, new UserCanceledException("User canceled Venmo."));
                 break;
             case BrowserSwitchStatus.SUCCESS:
@@ -494,14 +500,14 @@ public class VenmoClient {
                                             if (nonce != null) {
                                                 boolean shouldVault = sharedPrefsWriter.getVenmoVaultOption(context);
                                                 if (shouldVault && isClientTokenAuth) {
-                                                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success");
+                                                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success", payPalContextId, linkType);
                                                     vaultVenmoAccountNonce(nonce.getString(), callback);
                                                 } else {
-                                                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success");
+                                                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success", payPalContextId, linkType);
                                                     callback.onResult(nonce, null);
                                                 }
                                             } else {
-                                                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure");
+                                                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure", payPalContextId, linkType);
                                                 callback.onResult(null, error);
                                             }
                                         }
@@ -509,29 +515,29 @@ public class VenmoClient {
                                 } else if (paymentMethodNonce != null && username != null) {
                                     boolean shouldVault = sharedPrefsWriter.getVenmoVaultOption(context);
                                     if (shouldVault && isClientTokenAuth) {
-                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success");
+                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success", payPalContextId, linkType);
                                         vaultVenmoAccountNonce(paymentMethodNonce, callback);
                                     } else {
-                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success");
+                                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.success", payPalContextId, linkType);
                                         VenmoAccountNonce venmoAccountNonce = new VenmoAccountNonce(paymentMethodNonce, username, false);
                                         callback.onResult(venmoAccountNonce, null);
                                     }
                                 }
                             } else if (authError != null) {
-                                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure");
+                                braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure", payPalContextId, linkType);
                                 callback.onResult(null, authError);
                             }
                         }
                         });
                     } else if (deepLinkUri.getPath().contains("cancel")) {
-                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.canceled");
+                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.canceled", payPalContextId, linkType);
                         callback.onResult(null, new UserCanceledException("User canceled Venmo."));
                     } else if (deepLinkUri.getPath().contains("error")) {
-                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure");
+                        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure", payPalContextId, linkType);
                         callback.onResult(null, new Exception("Error returned from Venmo."));
                     }
                 } else {
-                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure");
+                    braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.failure", payPalContextId, linkType);
                     callback.onResult(null, new Exception("Unknown error"));
                 }
                 break;
@@ -697,7 +703,7 @@ public class VenmoClient {
                 .returnUrlScheme(braintreeClient.getReturnUrlScheme());
 
         braintreeClient.startBrowserSwitch(activity, browserSwitchOptions);
-        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.started");
+        braintreeClient.sendAnalyticsEvent("pay-with-venmo.app-links.started", payPalContextId, linkType);
     }
 
     /**