+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+
+* Venmo
+  * Send `link_type` in `event_params` to PayPal's analytics service (FPTI)
+  
 ## 4.45.0 (2024-04-16)
 
 * BraintreeCore